+++ src/com/ValidationSystem.java
@@ -1,12 +1,14 @@
 package com;
 
-class ValidationSystem {
-    static void validate(String input) throws ValidationFailedException {
+import com.exception.ValidationFailedException;
+
+public class ValidationSystem {
+    public static void validate(String input) throws ValidationFailedException {
         Validator<String> validator = new StringValidator();
         validator.validate(input);
     }
 
-    static void validate(Integer input) throws ValidationFailedException {
+    public static void validate(Integer input) throws ValidationFailedException {
         Validator<Integer> validator = new IntegerValidator();
         validator.validate(input);
     }
