+++ src/com/Validator.java
@@ -1,5 +1,7 @@
 package com;
 
+import com.exception.ValidationFailedException;
+
 public interface Validator<T> {
     void validate(T input) throws ValidationFailedException;
 }
