+++ src/com/exception/ValidationFailedException.java
@@ -1,4 +1,4 @@
-package com;
+package com.exception;
 
 public class ValidationFailedException extends Exception {
     public ValidationFailedException(String msg) {
